% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Datasets}
\alias{Datasets}
\title{Returns an object suitable for the Datasets argument of the bulk export
endpoints: \code{export_bulk} and \code{export_time_aligned}.}
\usage{
Datasets(Identifier, Calculation = NULL, Unit = NULL)
}
\description{
Returns an object suitable for the Datasets argument of the bulk export
endpoints: \code{export_bulk} and \code{export_time_aligned}.
}
