% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_row}
\alias{format_row}
\title{Helper function for simplifying common pattern in AQTS json responses}
\usage{
format_row(x)
}
\description{
Helper function for simplifying common pattern in AQTS json responses
}
