% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{webportal-routes}
\alias{webportal-routes}
\alias{version}
\alias{get_filters}
\alias{get_location}
\alias{get_locations}
\alias{get_dataset}
\alias{get_latest_stat}
\alias{get_latest_stat_values}
\alias{get_map_locations}
\alias{get_map_datasets}
\alias{get_map_latest_stat}
\alias{get_map_periodic_stat}
\alias{export_dataset}
\alias{export_periodic_stat}
\alias{export_seasonal_stat}
\alias{export_time_aligned}
\alias{export_bulk}
\title{Web Portal API}
\usage{
version(.perform = TRUE, .format = TRUE)

get_filters(.perform = TRUE, .format = TRUE)

get_location(location, ..., .format = TRUE, .perform = TRUE)

get_locations(location = NULL, active = NULL, .format = TRUE, .perform = TRUE)

get_dataset(dataset, .perform = TRUE, .format = TRUE)

get_latest_stat(
  parameter = NULL,
  statistic = NULL,
  active = NULL,
  .perform = TRUE,
  .format = TRUE
)

get_latest_stat_values(
  parameter = NULL,
  statistic = NULL,
  location = NULL,
  ...,
  .format = TRUE,
  .perform = TRUE
)

get_map_locations(..., .perform = TRUE, .format = TRUE)

get_map_datasets(parameter, ..., .perform = TRUE, .format = TRUE)

get_map_latest_stat(parameter, statistic, ..., .perform = TRUE, .format = TRUE)

get_map_periodic_stat(
  parameter,
  statistic,
  interval,
  date,
  ...,
  .perform = TRUE,
  .format = TRUE
)

export_dataset(dataset, ..., .perform = TRUE, .format = TRUE)

export_periodic_stat(
  dataset,
  statistic,
  calendar,
  interval,
  ...,
  .perform = TRUE,
  .format = TRUE
)

export_seasonal_stat(
  dataset,
  interval,
  statistic,
  referenceperiod = referenceperiod,
  ...,
  .perform = TRUE,
  .format = TRUE
)

export_time_aligned(
  datasets,
  starttime,
  endtime,
  daterange = NULL,
  interval = "PointsAsRecorded",
  step = 1,
  includeGradeCodes = TRUE,
  includeQualifers = TRUE,
  includeApprovalLevels = TRUE,
  ...,
  .perform = TRUE,
  .format = TRUE
)

export_bulk(
  datasets,
  starttime,
  endtime,
  daterange = NULL,
  interval = "PointsAsRecorded",
  step = 1,
  includeGradeCodes = TRUE,
  includeQualifers = TRUE,
  includeApprovalLevels = TRUE,
  ...,
  .perform = TRUE,
  .format = TRUE
)
}
\arguments{
\item{.perform}{if \code{TRUE}, perform the request, otherwise return a
\code{httr2::request} object}

\item{.format}{if \code{TRUE}, return a \code{data.frame}, otherwise return a
\code{httr2::response} object. Ignored if \code{.perform = FALSE}}

\item{...}{pass query arguments to route. Please see Web Portal API
documenation for available arguments.}
}
\description{
See the web portal API documentation for arguments.
}
\section{Functions}{
\itemize{
\item \code{version()}: Get version of API

\item \code{get_filters()}: Get filters for map endpoints

\item \code{get_location()}: Get a location

\item \code{get_locations()}: Get locations

\item \code{get_dataset()}: Get datasets

\item \code{get_latest_stat()}: Get Latest Statistics

\item \code{get_latest_stat_values()}: Get Latest Statistics

\item \code{get_map_locations()}: List all locations

\item \code{get_map_datasets()}: List timeseries by parameter; return sf object

\item \code{get_map_latest_stat()}: Get latest statistics; return sf object

\item \code{get_map_periodic_stat()}: Get periodic statistics; return sf object

\item \code{export_dataset()}: Export a data set

\item \code{export_periodic_stat()}: Export a periodic statistic

\item \code{export_seasonal_stat()}: Export seasonal statistics

\item \code{export_time_aligned()}: Export several time-aligned data sets.

\item \code{export_bulk()}: Export datasets in bulk; optionally summarize

}}
